+++ tests/test_dependencies_async.py
@@ -15,7 +15,7 @@ async def test_async_invoke(incanter: Incanter):
         await sleep(0.001)
         return 2
 
-    assert (await incanter.acall(fn)) == 2
+    assert (await incanter.acompose_and_call(fn)) == 2
 
 
 async def test_async_dep(incanter: Incanter):
@@ -24,9 +24,9 @@ async def test_async_dep(incanter: Incanter):
         return 1
 
     with pytest.raises(TypeError):
-        incanter.call(lambda dep1: dep1 + 1)
+        incanter.compose_and_call(lambda dep1: dep1 + 1)
 
-    assert (await incanter.acall(lambda dep1: dep1 + 1)) == 2
+    assert (await incanter.acompose_and_call(lambda dep1: dep1 + 1)) == 2
     assert signature(
         incanter.compose(lambda dep1: dep1 + 1, is_async=True)
     ).parameters == OrderedDict([])
@@ -58,9 +58,9 @@ async def test_async_mixed_dep(incanter: Incanter):
         return input + 1
 
     with pytest.raises(TypeError):
-        incanter.call(lambda dep1: dep1 + 1, 1)
+        incanter.compose_and_call(lambda dep1: dep1 + 1, 1)
 
-    assert (await incanter.acall(lambda dep1: dep1 + 1, 1)) == 4
+    assert (await incanter.acompose_and_call(lambda dep1: dep1 + 1, 1)) == 4
     assert signature(
         incanter.compose(lambda dep1: dep1 + 1, is_async=True)
     ).parameters == OrderedDict(
@@ -84,7 +84,7 @@ async def test_async_ctx_manager_dep(incanter: Incanter):
         nonlocal entered
         return dep1 + 1
 
-    assert (await incanter.acall(fn)) == 2
+    assert (await incanter.acompose_and_call(fn)) == 2
 
     assert entered
     assert exited
@@ -98,7 +98,7 @@ async def test_taskgroup_dep(incanter: Incanter):
         assert tg
         return 2
 
-    assert (await incanter.acall(fn)) == 2
+    assert (await incanter.acompose_and_call(fn)) == 2
 
 
 def test_async_invoke_return_type(incanter: Incanter):