+++ tests/test_params.py
@@ -10,7 +10,7 @@ def test_params_with_defaults():
     def fn(default_param=1):
         return default_param
 
-    assert incant.call(fn) == 1
+    assert incant.compose_and_call(fn) == 1
 
     assert signature(incant.compose(fn)) == Signature(
         [Parameter("default_param", Parameter.POSITIONAL_OR_KEYWORD, default=1)]