+++ tests/test_string_annotations.py
@@ -20,13 +20,13 @@ def test_simple_prepare(incanter: Incanter):
         return dep1 + 1
 
     with pytest.raises(TypeError):
-        incanter.call(func)
+        incanter.compose_and_call(func)
     assert signature(incanter.compose(func)).parameters == {
         "dep1": Parameter("dep1", Parameter.POSITIONAL_OR_KEYWORD)
     }
 
     incanter.register_hook(lambda p: p.name == "dep1", lambda: 2)
-    assert incanter.call(func) == 3
+    assert incanter.compose_and_call(func) == 3
 
     assert signature(incanter.compose(func)).parameters == {}
     assert signature(incanter.compose(func)).return_annotation is int
@@ -40,7 +40,7 @@ def test_reg_by_type(incanter: Incanter):
     def fn(arg: str):
         return "1.0" + arg
 
-    assert incanter.call(fn, 1.0) == "1.02.0"
+    assert incanter.compose_and_call(fn, 1.0) == "1.02.0"
 
 
 def test_incant_pos_args_by_type(incanter: Incanter):